refactor(chat): extract action-suggestions cluster from the god file
Moves the "Action Suggestions" cluster (clearSuggestions + generateActions —
the LLM-generated quick-reply chips) out of chat_service.dart into a
`part`/extension file, chat_service_actions.dart. Verbatim, zero behaviour
change; the cluster sits cleanly between two methods with the periodic-eval
counter field immediately after it, so the cut needed no field surgery.

Net effect: chat_service.dart drops 4,213 -> 4,093 lines (-120).

flutter analyze: clean. Full unit suite (1348) + the 1:1<->group realism parity
gate: green.
